added several tags to one tg user

+++ CrTelegramBot/Services/UserLinkService.cs
@@ -16,7 +16,7 @@ public sealed class UserLinkService
     }
 
     public Task<UserLink?> FindByTelegramUserIdAsync(long userId, CancellationToken ct = default) =>
-        _db.UserLinks.FirstOrDefaultAsync(x => x.TelegramUserId == userId, ct);
+        _db.UserLinks.OrderBy(x => x.Id).FirstOrDefaultAsync(x => x.TelegramUserId == userId, ct);
 
     public Task<UserLink?> FindByPlayerTagAsync(string playerTag, CancellationToken ct = default) =>
         _db.UserLinks.FirstOrDefaultAsync(x => x.PlayerTag == ClashRoyaleApiClient.NormalizeTag(playerTag), ct);
@@ -24,56 +24,74 @@ public sealed class UserLinkService
     public async Task<UserLink> UpsertAsync(User user, string playerTag, CancellationToken ct = default)
     {
         playerTag = ClashRoyaleApiClient.NormalizeTag(playerTag);
-        var byUser = await _db.UserLinks.FirstOrDefaultAsync(x => x.TelegramUserId == user.Id, ct);
-        
-        if (byUser is not null)
-            _db.UserLinks.Remove(byUser);
+        var displayName = string.Join(' ', new[] { user.FirstName, user.LastName }.Where(x => !string.IsNullOrWhiteSpace(x)));
+        var username = user.Username ?? string.Empty;
 
         var byTag = await _db.UserLinks.FirstOrDefaultAsync(x => x.PlayerTag == playerTag, ct);
-        
+
         if (byTag is not null)
+        {
+            if (byTag.TelegramUserId == user.Id)
+            {
+                byTag.TelegramUsername = username;
+                byTag.TelegramDisplayName = displayName;
+                byTag.IsEnabled = true;
+                await _db.SaveChangesAsync(ct);
+                return byTag;
+            }
+
             _db.UserLinks.Remove(byTag);
+        }
 
         var link = new UserLink
         {
             TelegramUserId = user.Id,
-            TelegramUsername = user.Username ?? string.Empty,
-            TelegramDisplayName = string.Join(' ', new[] { user.FirstName, user.LastName }.Where(x => !string.IsNullOrWhiteSpace(x))),
+            TelegramUsername = username,
+            TelegramDisplayName = displayName,
             PlayerTag = playerTag,
             IsEnabled = true
         };
 
         _db.UserLinks.Add(link);
         await _db.SaveChangesAsync(ct);
-        
+
         return link;
     }
 
     public async Task<UserLink> UpsertAsync(long telegramUserId, string telegramUsername, string telegramDisplayName, string playerTag, CancellationToken ct = default)
     {
         playerTag = ClashRoyaleApiClient.NormalizeTag(playerTag);
-        var byUser = await _db.UserLinks.FirstOrDefaultAsync(x => x.TelegramUserId == telegramUserId, ct);
-        
-        if (byUser is not null)
-            _db.UserLinks.Remove(byUser);
+        var username = telegramUsername ?? string.Empty;
+        var displayName = telegramDisplayName ?? string.Empty;
 
         var byTag = await _db.UserLinks.FirstOrDefaultAsync(x => x.PlayerTag == playerTag, ct);
-        
+
         if (byTag is not null)
+        {
+            if (byTag.TelegramUserId == telegramUserId)
+            {
+                byTag.TelegramUsername = username;
+                byTag.TelegramDisplayName = displayName;
+                byTag.IsEnabled = true;
+                await _db.SaveChangesAsync(ct);
+                return byTag;
+            }
+
             _db.UserLinks.Remove(byTag);
+        }
 
         var link = new UserLink
         {
             TelegramUserId = telegramUserId,
-            TelegramUsername = telegramUsername ?? string.Empty,
-            TelegramDisplayName = telegramDisplayName ?? string.Empty,
+            TelegramUsername = username,
+            TelegramDisplayName = displayName,
             PlayerTag = playerTag,
             IsEnabled = true
         };
 
         _db.UserLinks.Add(link);
         await _db.SaveChangesAsync(ct);
-        
+
         return link;
     }
 
@@ -81,7 +99,7 @@ public sealed class UserLinkService
     {
         _db.UserLinks.Remove(link);
         await _db.SaveChangesAsync(ct);
-        
+
         return true;
     }
 }
